+++ src/components/config/config.test.tsx
@@ -1,6 +1,7 @@
-import React from 'react';
-import { config, setLocationSrv } from '@grafana/runtime';
 import { shallow } from 'enzyme';
+import React from 'react';
+import { setLocationSrv } from '@grafana/runtime';
+import { ApplicationRoot } from '../../constants';
 import { Config } from './config';
 
 /*
@@ -18,8 +19,6 @@ const getPlugin = (overridePlugin: any = { meta: {} }) => ({
  Config
  */
 describe('Config', () => {
-  let initialDataSources = config.datasources;
-
   beforeAll(() => {
     jest.spyOn(Config, 'getLocation').mockImplementation((): any => ({
       assign: jest.fn(),
@@ -31,36 +30,22 @@ describe('Config', () => {
    Initialization
    */
   describe('Initialization', () => {
-    it('If plugin is not enabled, state should have isEnabled = false and isConfigured = false', () => {
+    it('If plugin is not enabled and meta is not set, state should have isEnabled = false', () => {
+      const plugin = getPlugin({});
+      const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
+      expect(wrapper.state().isEnabled).toBeTruthy();
+    });
+
+    it('If plugin is not enabled, state should have isEnabled = false', () => {
       const plugin = getPlugin({ meta: { enabled: false } });
-      config.datasources = {
-        redis: {
-          type: 'redis-datasource',
-        } as any,
-      };
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
       expect(wrapper.state().isEnabled).toBeFalsy();
-      expect(wrapper.state().isConfigured).toBeFalsy();
     });
 
-    it('If plugin is enabled but config does not have needed datasources, state should have isEnabled = true and isConfigured = false', () => {
+    it('If plugin is enabled, state should have isEnabled = true', () => {
       const plugin = getPlugin({ meta: { enabled: true } });
-      config.datasources = {};
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
       expect(wrapper.state().isEnabled).toBeTruthy();
-      expect(wrapper.state().isConfigured).toBeFalsy();
-    });
-
-    it('If plugin is enabled and config has needed datasources, state should have isEnabled = true and isConfigured = true', () => {
-      const plugin = getPlugin({ meta: { enabled: true } });
-      config.datasources = {
-        redis: {
-          type: 'redis-datasource',
-        } as any,
-      };
-      const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
-      expect(wrapper.state().isEnabled).toBeTruthy();
-      expect(wrapper.state().isConfigured).toBeTruthy();
     });
   });
 
@@ -68,30 +53,18 @@ describe('Config', () => {
    Rendering
    */
   describe('rendering', () => {
-    beforeAll(() => {
-      config.datasources = {
-        redis: {
-          type: 'redis-datasource',
-        } as any,
-      };
-    });
-
-    it('If plugin is not configured, should show only enable button', () => {
+    it('If plugin is not configured, should show enable button', () => {
       const plugin = getPlugin({ meta: { enabled: false } });
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
       const enableButton = wrapper.findWhere((node) => node.name() === 'Button' && node.text() === 'Enable');
-      const updateButton = wrapper.findWhere((node) => node.name() === 'Button' && node.text() === 'Update');
       expect(enableButton.exists()).toBeTruthy();
-      expect(updateButton.exists()).not.toBeTruthy();
     });
 
-    it('If plugin is configured, should show update and disable buttons', () => {
+    it('If plugin is configured, should show disable buttons', () => {
       const plugin = getPlugin({ meta: { enabled: true } });
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
       const disableButton = wrapper.findWhere((node) => node.name() === 'Button' && node.text() === 'Disable');
-      const updateButton = wrapper.findWhere((node) => node.name() === 'Button' && node.text() === 'Update');
       expect(disableButton.exists()).toBeTruthy();
-      expect(updateButton.exists()).toBeTruthy();
     });
 
     it('Enable button should call onEnable method', () => {
@@ -113,38 +86,12 @@ describe('Config', () => {
       disableButton.simulate('click');
       expect(testedMethod).toHaveBeenCalled();
     });
-
-    it('Update button should call onUpdate method', () => {
-      const plugin = getPlugin({ meta: { enabled: true } });
-      const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
-      const testedMethod = jest.spyOn(wrapper.instance(), 'onUpdate').mockImplementation(() => null);
-      wrapper.instance().forceUpdate();
-      const updateButton = wrapper.findWhere((node) => node.name() === 'Button' && node.text() === 'Update');
-      updateButton.simulate('click');
-      expect(testedMethod).toHaveBeenCalled();
-    });
   });
 
   /*
    Methods
    */
   describe('Methods', () => {
-    it('onUpdate should call goHome method', () => {
-      const plugin = getPlugin({ meta: { enabled: true } });
-      const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
-      const testedMethod = jest.spyOn(wrapper.instance(), 'goHome').mockImplementation();
-      wrapper.instance().onUpdate();
-      expect(testedMethod).toHaveBeenCalled();
-    });
-
-    it('onUpdate should not call goHome method if enabled=false', () => {
-      const plugin = getPlugin({ meta: { enabled: false } });
-      const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
-      const testedMethod = jest.spyOn(wrapper.instance(), 'goHome').mockImplementation();
-      wrapper.instance().onUpdate();
-      expect(testedMethod).not.toHaveBeenCalled();
-    });
-
     it('onDisable should call updatePluginSettings method', () => {
       const plugin = getPlugin({ meta: { enabled: true } });
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
@@ -186,14 +133,13 @@ describe('Config', () => {
       const wrapper = shallow<Config>(<Config plugin={plugin} query={null as any} />);
       wrapper.instance().goHome();
       expect(updateLocationMock).toHaveBeenCalledWith({
-        path: 'a/redis-app/',
+        path: ApplicationRoot,
         partial: false,
       });
     });
   });
 
   afterAll(() => {
-    config.datasources = initialDataSources;
     jest.resetAllMocks();
   });
 });