+++ src/redis-latency-panel/components/redis-latency-panel-graph/redis-latency-panel-graph.test.tsx
@@ -5,6 +5,13 @@ import { DefaultColorModeId } from '../../constants';
 import { RedisLatencyPanelGraph } from './redis-latency-panel-graph';
 
 /**
+ * Mock @grafana/runtime
+ */
+jest.mock('@grafana/runtime', () => ({
+  config: { theme2: {} },
+}));
+
+/**
  * Latency Panel Table
  */
 describe('RedisLatencyPanelGraph', () => {