+++ src/redis-latency-panel/components/redis-latency-panel-table/redis-latency-panel-table.test.tsx
@@ -6,6 +6,13 @@ import { DisplayNameByFieldName, FieldName } from '../../constants';
 import { RedisLatencyPanelTable } from './redis-latency-panel-table';
 
 /**
+ * Mock @grafana/runtime
+ */
+jest.mock('@grafana/runtime', () => ({
+  config: { theme2: {} },
+}));
+
+/**
  * Latency Panel Table
  */
 describe('RedisLatencyPanel', () => {