+++ CHANGELOG.md
@@ -1,12 +1,31 @@
 # Change Log
 
-## 1.2.0 (IN PROGRESS)
+## 2.0.0 (IN PROGRESS)
+
+### Bug fixes
+
+- "Available Requirements" panel should be set to $redis datasource #63
+- Cannot read property 'v1' of undefined (theme.v1) in the Grafana8 #65
+- Upgrade to Grafana 8.0.2 (#67)
+
+## 1.2.0 (2021-05-11)
 
 ### Features / Enhancements
 
-- Update Grafana dependencies to 7.5.4
+- Upgrade Grafana dependencies 7.5.4 #53
+- Update Docker workflow #52
+- Update Docker token and add Master build #54
+- Add RefId to Query as mandatory for the upcoming release #55
+- Update RedisGears Script Editor Execution modes #57
+- Update Dashboards to version 7.5.5 #58
 
-## v1.1.0 (2021-02-07)
+### Bug fixes
+
+- Add default Color Mode Id for older releases (7.2.X) #49
+- Fix Latency below zero calculation #56
+- "Cannot read property 'Tooltip' of undefined" for Latency Panel in the upcoming release #60
+
+## 1.1.0 (2021-02-07)
 
 ### Features / Enhancements
 
@@ -26,14 +45,14 @@
 - Add Docker build #42
 - Update Plugin and panels configuration for Redis Data Source 1.3.1 #44
 
-## v1.0.1 (2020-10-24)
+## 1.0.1 (2020-10-24)
 
 ### Features / Enhancements
 
 - Add GitHub action to sign release #13
 - Signed release
 
-## v1.0.0 (2020-10-14)
+## 1.0.0 (2020-10-14)
 
 ### Features / Enhancements